requirements.txt
Appium-Python-Client==5.1.1
configparser